Update symlink for files

+++ dotfiles/createsymlinks.sh
@@ -8,7 +8,7 @@
 
 dir=~/fresh_install/dotfiles                    # dotfiles directory
 olddir=~/dotfiles_old             # old dotfiles backup directory
-files="vimrc"    # list of files/folders to symlink in homedir
+declare -a files=("vimrc" "zshrc" "tmux.conf" "tmux.conf.local")    # list of files/folders to symlink in homedir
 
 ##########
 
@@ -23,9 +23,12 @@ cd $dir
 echo "...done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
-for file in $files; do
-	    echo "Moving any existing dotfiles from ~ to $olddir"
-	        mv ~/.$file ~/dotfiles_old/
-		    echo "Creating symlink to $file in home directory."
-		        ln -s $dir/$file ~/.$file
-		done
+for file in "${files[@]}"; do
+    echo $file
+    if [ -e $file ]; then
+        echo "Moving any existing dotfiles from ~ to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+    fi
+    echo "Creating symlink to $file in home directory."
+    ln -s $dir/$file ~/.$file		
+done
